+++ WpfAutoGitHelper/ViewModels/AppDialogMode.cs
@@ -1,0 +1,9 @@
+namespace WpfAutoGitHelper.ViewModels
+{
+    public enum AppDialogMode
+    {
+        Ok,
+        YesNo,
+        InputOkCancel,
+    }
+}